+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -18,6 +18,7 @@
 - Diagnostics lost during JSON serialization
 - Files with extension should not be set to hlasm in libs folder
 - Lookahead mode does not work correctly when triggered from AINSERTed code
+- Remove ALIAS operand parsing limitation
 
 ## [0.14.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/0.13.0...0.14.0) (2021-08-18)